MoscariniPostelVinay2016

Documentation for MoscariniPostelVinay2016.

MoscariniPostelVinay2016.plot_final

MoscariniPostelVinay2016.simvalues

MoscariniPostelVinay2016.plot_final — Method

plot_final(SIM, S, P, EXO)

Reproduce the plots from the paper to check our results: https://www.aeaweb.org/articles?id=10.1257/aer.p20161051 ...

Arguments

SIM::Dict: Output from wagepost_gyn.jl

S::Dict: Output from calibration.jl

P::Dict: Output from calibration.jl

EXO::Dict: Output from wagepost_gyn.jl

...

MoscariniPostelVinay2016.simvalues — Method

Returning the structure SIM of all simulated series, including an updated path of job ads at(p), given an initial guess a0t(p) and a realization of aggregate states. ...

Arguments

om::int: omega

S::Dict: Output from calibration.jl

TB::Dict: Output from calibration.jl

exit_chk::string: Check for wether we allow for exit or not

a0_in::Vector{Float64}: sampling distribution for current guess of a(p)

x::Vector{Float64}: initial guess for x=[0, 0.005, 0.07]

P::Dict: Output from calibration.jl

c_st_i::Matrix{Int64}: aggregate states

a0_type::string: a0_in type

...